Fix bug when toggling the timer, countup start from 0
This is because `slide_countdown` doesn't support counting from N to
infinity, you can either count from 0 to infinity or from 0 to N, N here
is the `duration` argument. So here the issue was fixed by implementing
counting up to be always from N to infinity, we don't need other modes
for now.

This is a quick fix until the PR in the main `slide_countdown` project is merged.